refactor(hydro_lang)!: move `ir` module under `builder`
Most Hydro users don't need to interact with the `ir`, and its primary consumer is the compilation logic inside `builder`.
