+++ hydro_lang/src/builder/deploy.rs
@@ -16,7 +16,7 @@ use super::compiled::CompiledFlow;
 use super::deploy_provider::{
     ClusterSpec, Deploy, ExternalSpec, IntoProcessSpec, Node, ProcessSpec, RegisterPort,
 };
-use crate::ir::HydroRoot;
+use super::ir::HydroRoot;
 use crate::location::external_process::{
     ExternalBincodeBidi, ExternalBincodeSink, ExternalBincodeStream, ExternalBytesPort,
 };
