+++ hydro_lang/src/graph/api.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 
+use crate::builder::ir::HydroRoot;
 use crate::graph::render::HydroWriteConfig;
-use crate::ir::HydroRoot;
 
 /// Graph generation API for built flows
 pub struct GraphApi<'a> {
