+++ hydro_lang/src/lib.rs
@@ -86,9 +86,6 @@ pub mod builder;
 
 mod manual_expr;
 
-#[expect(missing_docs, reason = "TODO")]
-pub mod ir;
-
 #[cfg(feature = "viz")]
 #[expect(missing_docs, reason = "TODO")]
 pub mod graph;
