+++ hydro_lang/src/live_collections/keyed_stream.rs
@@ -7,8 +7,8 @@ use super::keyed_singleton::KeyedSingleton;
 use super::optional::Optional;
 use super::stream::{ExactlyOnce, MinOrder, MinRetries, NoOrder, Stream, TotalOrder};
 use crate::boundedness::{Bounded, Boundedness, Unbounded};
+use crate::builder::ir::HydroNode;
 use crate::cycle::{CycleCollection, CycleComplete, ForwardRefMarker};
-use crate::ir::HydroNode;
 use crate::location::tick::NoAtomic;
 use crate::location::{Atomic, Location, LocationId, NoTick, Tick, check_matching_location};
 use crate::manual_expr::ManualExpr;
