+++ hydro_lang/src/location/mod.rs
@@ -12,12 +12,12 @@ use syn::parse_quote;
 use tokio_util::codec::{Decoder, Encoder, LengthDelimitedCodec};
 
 use super::builder::FlowState;
-use crate::ir::backtrace::Backtrace;
 use crate::boundedness::Unbounded;
-use crate::cycle::{CycleCollection, ForwardRef, ForwardRefMarker};
-use crate::ir::{
+use crate::builder::ir::backtrace::Backtrace;
+use crate::builder::ir::{
     DebugInstantiate, HydroIrMetadata, HydroIrOpMetadata, HydroNode, HydroRoot, HydroSource,
 };
+use crate::cycle::{CycleCollection, ForwardRef, ForwardRefMarker};
 use crate::live_collections::keyed_stream::KeyedStream;
 use crate::live_collections::singleton::Singleton;
 use crate::live_collections::stream::{ExactlyOnce, NoOrder, Stream, TotalOrder};
