+++ hydro_lang/src/location/tick.rs
@@ -7,11 +7,11 @@ use stageleft::{QuotedWithContext, q};
 use super::{Cluster, Location, LocationId, Process};
 use crate::boundedness::Bounded;
 use crate::builder::FlowState;
+use crate::builder::ir::{HydroNode, HydroSource};
 use crate::cycle::{
     CycleCollection, CycleCollectionWithInitial, DeferTick, ForwardRef, ForwardRefMarker,
     TickCycle, TickCycleMarker,
 };
-use crate::ir::{HydroNode, HydroSource};
 use crate::live_collections::optional::Optional;
 use crate::live_collections::singleton::Singleton;
 use crate::live_collections::stream::{ExactlyOnce, Stream, TotalOrder};
