+++ hydro_lang/src/rewrites/properties.rs
@@ -2,7 +2,7 @@ use std::collections::HashSet;
 
 use stageleft::*;
 
-use crate::ir::{HydroNode, HydroRoot, transform_bottom_up};
+use crate::builder::ir::{HydroNode, HydroRoot, transform_bottom_up};
 
 /// Structure for tracking expressions known to have particular algebraic properties.
 ///
