+++ hydro_optimize/src/deploy_and_analyze.rs
@@ -3,11 +3,11 @@ use std::time::Duration;
 
 use hydro_deploy::Deployment;
 use hydro_lang::builder::deploy::DeployResult;
+use hydro_lang::builder::ir::{HydroNode, HydroRoot, deep_clone, traverse_dfir};
 use hydro_lang::builder::{FlowBuilder, RewriteIrFlowBuilder};
 use hydro_lang::deploy::HydroDeploy;
 use hydro_lang::deploy::deploy_graph::DeployCrateWrapper;
 use hydro_lang::internal_constants::{COUNTER_PREFIX, CPU_USAGE_PREFIX};
-use hydro_lang::ir::{HydroNode, HydroRoot, deep_clone, traverse_dfir};
 use hydro_lang::location::LocationId;
 use hydro_lang::rewrites::persist_pullup::persist_pullup;
 use tokio::sync::mpsc::UnboundedReceiver;
