+++ hydro_optimize/src/parse_results.rs
@@ -1,9 +1,9 @@
 use std::collections::HashMap;
 
 use hydro_lang::builder::deploy::DeployResult;
+use hydro_lang::builder::ir::{HydroNode, HydroRoot, traverse_dfir};
 use hydro_lang::deploy::HydroDeploy;
 use hydro_lang::deploy::deploy_graph::DeployCrateWrapper;
-use hydro_lang::ir::{HydroNode, HydroRoot, traverse_dfir};
 use hydro_lang::location::LocationId;
 use regex::Regex;
 use tokio::sync::mpsc::UnboundedReceiver;
