+++ hydro_optimize/src/partition_syn_analysis.rs
@@ -779,8 +779,8 @@ mod tests {
     use std::collections::BTreeMap;
 
     use hydro_lang::builder::FlowBuilder;
+    use hydro_lang::builder::ir::{HydroNode, HydroRoot, deep_clone, traverse_dfir};
     use hydro_lang::deploy::HydroDeploy;
-    use hydro_lang::ir::{HydroNode, HydroRoot, deep_clone, traverse_dfir};
     use hydro_lang::location::Location;
     use stageleft::q;
     use syn::visit::Visit;
