+++ hydro_optimize/src/repair.rs
@@ -1,7 +1,7 @@
 use std::cell::RefCell;
 use std::collections::HashMap;
 
-use hydro_lang::ir::{HydroNode, HydroRoot, transform_bottom_up, traverse_dfir};
+use hydro_lang::builder::ir::{HydroNode, HydroRoot, transform_bottom_up, traverse_dfir};
 use hydro_lang::location::LocationId;
 use syn::Ident;
